fix/Types refactoring pogo types and chips data source
