+++ lib/pages/home.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:pogo_diary/blocs/bloc_provider.dart';
 import 'package:pogo_diary/blocs/pokemon_type_chips_bloc.dart';
 import 'package:pogo_diary/blocs/pokemon_type_info_bloc.dart';
-import 'package:pogo_diary/data/pokemon_type_value.dart';
+import 'package:pogo_diary/data/pokemon_type_chips.dart';
 import 'package:pogo_diary/pages/pokemon_type_info.dart';
 import 'package:pogo_diary/widgets/pokemon_type_chip.dart';
